Latest changes from laptop

+++ .zshrc
@@ -54,12 +54,9 @@ plugins=(git brew dircycle osx python screen tmux)
 
 # User configuration
 
-export PATH="$HOME/src/git-external/bin:$HOME/local/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin"
-# export PYTHONPATH="$HOME/code/selenium-all/src/acs-llc-interface/vehiclecaninterface/scripts:$HOME/code/selenium-all/build/datatypes:$HOME/code/selenium-all/build/acsllcinterface_datatypes_python:/usr/local/lib/python2.7/site-packages"
-# So that Netbeans can find lldb debug server (a version that is correctly signed)
-export LLDB_DEBUGSERVER_PATH=/Applications/Xcode.app/Contents/SharedFrameworks/LLDB.framework/Resources/debugserver
-export JAVA_HOME=`/usr/libexec/java_home`
-export CODE_DIR="${HOME}/code"
+export PATH="$HOME/local/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin"
+export PATH="$HOME/src/git-external/bin:$PATH"
+export PATH="$HOME/.cargo/bin:$PATH"
 
 source $ZSH/oh-my-zsh.sh
 
@@ -92,13 +89,22 @@ unsetopt share_history
 # Example aliases
 # alias zshconfig="mate ~/.zshrc"
 # alias ohmyzsh="mate ~/.oh-my-zsh"
-alias netbeans="open /Applications/NetBeans/NetBeans\ 8.1.app & disown"
-alias clion="open /Applications/CLion.app & disown"
-alias qt-creator="open /Users/cswetenham/Qt/Qt\ Creator.app & disown"
-alias nuke-build="rm -rf build/ && mkdir build && cd build && cmake ../src -DCMAKE_BUILD_TYPE=RelWithDebInfo -DCMAKE_CXX_FLAGS=\"-Wall -Werror\" '-GCodeBlocks - Ninja' && ninja -j 10"
+# Config specific to work laptop
+if [[ $(hostname) == "chris-mbp" ]]; then
+  alias netbeans="open /Applications/NetBeans/NetBeans\ 8.1.app & disown"
+  alias clion="open /Applications/CLion.app & disown"
+  alias qt-creator="open /Users/cswetenham/Qt/Qt\ Creator.app & disown"
+  alias nuke-build="rm -rf build/ && mkdir build && cd build && cmake ../src -DCMAKE_BUILD_TYPE=RelWithDebInfo -DCMAKE_CXX_FLAGS=\"-Wall -Werror\" '-GCodeBlocks - Ninja' && ninja -j 10"
 
-eval $(ssh-agent)
+  eval $(ssh-agent)
 
-# Added by tools-python installer
-alias activate-oxpy='export MAGICK_HOME=/usr/local/Cellar/imagemagick@6/6.9.9-26 && source /Users/cswetenham/miniconda/bin/activate oxbotica'
-alias deactivate-oxpy='source /Users/cswetenham/miniconda/bin/deactivate oxbotica'
+  # Added by tools-python installer
+  alias activate-oxpy='export MAGICK_HOME=/usr/local/Cellar/imagemagick@6/6.9.9-26 && source /Users/cswetenham/miniconda/bin/activate oxbotica'
+  alias deactivate-oxpy='source /Users/cswetenham/miniconda/bin/deactivate oxbotica'
+
+  # export PYTHONPATH="$HOME/code/selenium-all/src/acs-llc-interface/vehiclecaninterface/scripts:$HOME/code/selenium-all/build/datatypes:$HOME/code/selenium-all/build/acsllcinterface_datatypes_python:/usr/local/lib/python2.7/site-packages"
+  # So that Netbeans can find lldb debug server (a version that is correctly signed)
+  export LLDB_DEBUGSERVER_PATH=/Applications/Xcode.app/Contents/SharedFrameworks/LLDB.framework/Resources/debugserver
+  export JAVA_HOME=`/usr/libexec/java_home`
+  export CODE_DIR="${HOME}/code"
+fi
